+++ actionpack/lib/action_dispatch/caching.rb
@@ -20,7 +20,6 @@ module ActionDispatch
         end
     end
 
-    include AbstractController::Helpers
     include ConfigMethods
     include ActionDispatch::Caching::Fragments
 
@@ -50,12 +49,12 @@ module ActionDispatch
 
     protected
       # Convenience accessor.
-    def cache(key, options = {}, &block)
-      if cache_configured?
-        cache_store.fetch(ActiveSupport::Cache.expand_cache_key(key, :controller), options, &block)
-      else
-        yield
+      def cache(key, options = {}, &block)
+        if cache_configured?
+          cache_store.fetch(ActiveSupport::Cache.expand_cache_key(key, :controller), options, &block)
+        else
+          yield
+        end
       end
-    end
   end
 end
